Move jacoco configuration to its own build script
